# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TestGyver API",
    "description": "API pour la plateforme de test multi-environnements TestGyver - Really, are you serious?",
    "version": "1.0.0",
    "contact": {
      "name": "TestGyver Support"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
    }
  },
...
  "paths": {
...
    "/tests": {
      "get": {
        "tags": ["Tests"],
        "summary": "Liste des tests",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "query",
            "name": "campain_id",
            "type": "string",
            "description": "Filtrer par campagne"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des tests",
            "schema": {
              "type": "array",
              "items": {"$ref": "#/definitions/Test"}
            }
          }
        }
      },
      "post": {
        "tags": ["Tests"],
        "summary": "Créer un test",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "body",
            "name": "test",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TestCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Test créé",
            "schema": {
              "$ref": "#/definitions/Test"
            }
          },
          "400": {"description": "Données invalides"}
        }
      }
    },
    "/tests/{test_id}": {
      "get": {
        "tags": ["Tests"],
        "summary": "Détails d'un test",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du test",
            "schema": {
              "$ref": "#/definitions/Test"
            }
          },
          "404": {"description": "Test non trouvé"}
        }
      },
      "put": {
        "tags": ["Tests"],
        "summary": "Mettre à jour un test",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "test",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TestUpdate"
            }
          }
        ],
        "responses": {
          "200": {"description": "Test mis à jour"},
          "404": {"description": "Test non trouvé"}
        }
      },
      "delete": {
        "tags": ["Tests"],
        "summary": "Supprimer un test",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {"description": "Test supprimé"},
          "404": {"description": "Test non trouvé"}
        }
      }
    },
    "/tests/{test_id}/actions": {
      "post": {
        "tags": ["Tests"],
        "summary": "Ajouter une action à un test",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "action",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Action"
            }
          }
        ],
        "responses": {
          "201": {"description": "Action ajoutée"},
          "404": {"description": "Test non trouvé"}
        }
      }
    },
    "/tests/filieres": {
      "get": {
        "tags": ["Tests"],
        "summary": "Liste des filières disponibles",
        "security": [{"Bearer": []}],
        "responses": {
          "200": {
            "description": "Liste des filières",
            "schema": {
              "type": "array",
              "items": {"type": "string"}
            }
          }
        }
      }
    },
    "/tests/{test_id}/execute": {
      "post": {
        "tags": ["Tests"],
        "summary": "Exécuter un test",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "execution",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "filiere": {"type": "string", "description": "Environnement d'exécution"}
              },
              "required": ["filiere"]
            }
          }
        ],
        "responses": {
          "200": {"description": "Exécution lancée"},
          "400": {"description": "Filière manquante"},
          "404": {"description": "Test non trouvé"}
        }
      }
    },
    "/tests/{test_id}/move-up": {
      "post": {
        "tags": ["Tests"],
        "summary": "Monter un test dans l'ordre d'exécution",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {"description": "Test déplacé"},
          "404": {"description": "Test non trouvé"}
        }
      }
    },
    "/tests/{test_id}/move-down": {
      "post": {
        "tags": ["Tests"],
        "summary": "Descendre un test dans l'ordre d'exécution",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {"description": "Test déplacé"},
          "404": {"description": "Test non trouvé"}
        }
      }
    },
    "/variables": {
      "get": {
        "tags": ["Variables"],
        "summary": "Liste des variables",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "query",
            "name": "grouped",
            "type": "boolean",
            "description": "Grouper par filière"
          },
          {
            "in": "query",
            "name": "filiere",
            "type": "string",
            "description": "Filtrer par filière"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des variables",
            "schema": {
              "type": "array",
              "items": {"$ref": "#/definitions/Variable"}
            }
          }
        }
      },
      "post": {
        "tags": ["Variables"],
        "summary": "Créer une variable",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "body",
            "name": "variable",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VariableCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Variable créée",
            "schema": {
              "$ref": "#/definitions/Variable"
            }
          },
          "400": {"description": "Données invalides"}
        }
      }
    },
    "/variables/{variable_id}": {
      "get": {
        "tags": ["Variables"],
        "summary": "Détails d'une variable",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "variable_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de la variable",
            "schema": {
              "$ref": "#/definitions/Variable"
            }
          },
          "404": {"description": "Variable non trouvée"}
        }
      },
      "put": {
        "tags": ["Variables"],
        "summary": "Mettre à jour une variable",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "variable_id",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "variable",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VariableUpdate"
            }
          }
        ],
        "responses": {
          "200": {"description": "Variable mise à jour"},
          "404": {"description": "Variable non trouvée"}
        }
      },
      "delete": {
        "tags": ["Variables"],
        "summary": "Supprimer une variable",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "path",
            "name": "variable_id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {"description": "Variable supprimée"},
          "404": {"description": "Variable non trouvée"}
        }
      }
    },
...
    "/deleted/tests": {
      "get": {
        "tags": ["Éléments supprimés"],
        "summary": "Liste des tests supprimés logiquement",
        "description": "Récupère tous les tests supprimés logiquement (admin uniquement)",
        "security": [{"Bearer": []}],
        "responses": {
          "200": {
            "description": "Liste des tests supprimés",
            "schema": {
              "type": "object",
              "properties": {
                "tests": {
                  "type": "array",
                  "items": {"$ref": "#/definitions/Test"}
                }
              }
            }
          },
          "401": {"description": "Non authentifié"},
          "403": {"description": "Accès interdit (admin requis)"}
        }
      }
    },
...
  },
  "definitions": {
...
    "Test": {
      "type": "object",
      "properties": {
        "_id": {"type": "string"},
        "campainId": {"type": "string"},
        "userId": {"type": "string"},
        "dateCreated": {"type": "string", "format": "date-time"},
        "name": {"type": "string"},
        "description": {"type": "string"},
        "order": {"type": "integer"},
        "variables": {
          "type": "array",
          "items": {"type": "string"}
        },
        "actions": {
          "type": "array",
          "items": {"$ref": "#/definitions/Action"}
        }
      }
    },
    "TestCreate": {
      "type": "object",
      "required": ["campainId", "name"],
      "properties": {
        "campainId": {"type": "string"},
        "name": {"type": "string"},
        "description": {"type": "string"},
        "variables": {
          "type": "array",
          "items": {"type": "string"}
        },
        "actions": {
          "type": "array",
          "items": {"$ref": "#/definitions/Action"}
        }
      }
    },
    "TestUpdate": {
      "type": "object",
      "properties": {
        "name": {"type": "string"},
        "description": {"type": "string"},
        "variables": {
          "type": "array",
          "items": {"type": "string"}
        },
        "actions": {
          "type": "array",
          "items": {"$ref": "#/definitions/Action"}
        }
      }
    },
    "Action": {
      "type": "object",
      "required": ["type", "value"],
      "properties": {
        "type": {"type": "string", "description": "Type d'action (http, ftp, sftp, ssh, webdav, io, var)"},
        "value": {"type": "object", "description": "Configuration spécifique à l'action"},
        "outputVariables": {
          "type": "object",
          "description": "Variables de sortie associées",
          "additionalProperties": {"type": "string"}
        }
      }
    },
...
    "Variable": {
      "type": "object",
      "properties": {
        "_id": {"type": "string"},
        "key": {"type": "string"},
        "value": {},
        "filiere": {"type": "string"},
        "description": {"type": "string"},
        "isRoot": {"type": "boolean", "default": false}
      }
    },
    "VariableCreate": {
      "type": "object",
      "required": ["key", "value"],
      "properties": {
        "key": {"type": "string"},
        "value": {},
        "filiere": {"type": "string"},
        "description": {"type": "string"},
        "isRoot": {"type": "boolean"}
      }
    },
    "VariableUpdate": {
      "type": "object",
      "properties": {
        "key": {"type": "string"},
        "value": {},
        "filiere": {"type": "string"},
        "description": {"type": "string"},
        "isRoot": {"type": "boolean"}
      }
    },
...
  }
}